requirements-dev.txt
pytest>=2.9.2
pytest-cover>=3.0.0
pytest-aiohttp==0.1.2
jsonschema>=2.5.1
